requirements.txt
Flask==2.3.3
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
langdetect==1.0.9
pygments==2.16.1
autopep8==2.0.4
flake8==6.1.0
mypy==1.7.0
black==23.10.1
isort==5.12.0
bandit==1.7.5
radon==6.0.1
pylint==3.0.2
ast-decompiler==0.7.0
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.1
tree-sitter-java==0.20.2
tree-sitter-cpp==0.20.0
colorama==0.4.6
rich==13.7.0